tests: move run_as_lxc to tests/tunneldigger.py
Use tunneldigger.py as library for multiple tests.
